requirements.txt
flask
flask-cors
python-dotenv
google
google-api-python-client
google-genai
google-auth
google-auth-oauthlib
google-auth-httplib2
Pillow
flask-cors
firebase-admin
requests
gunicorn
moviepy
reportlab

duckduckgo-search
google-cloud-aiplatform

